moved around declarations
