+++ src/evo/file.h
@@ -34,37 +34,7 @@ using namespace std;
 // TODO: What include is required for this?
 #define CLIENT_VERSION 1
 
-boost::filesystem::path GetDataDirectory()
-{
-    string strDataDir2 = GetArg("-datadir2", "");
-    if(strDataDir2 != "") return strDataDir2;
-
-    namespace fs = boost::filesystem;
-    // Windows < Vista: C:\Documents and Settings\Username\Application Data\Dash
-    // Windows >= Vista: C:\Users\Username\AppData\Roaming\Dash
-    // Mac: ~/Library/Application Support/Dash
-    // Unix: ~/.dash
-#ifdef WIN32
-    // Windows
-    return GetSpecialFolderPath(CSIDL_APPDATA) / "Dash";
-#else
-    fs::path pathRet;
-    char* pszHome = getenv("HOME");
-    if (pszHome == NULL || strlen(pszHome) == 0)
-        pathRet = fs::path("/");
-    else
-        pathRet = fs::path(pszHome);
-#ifdef MAC_OSX
-    // Mac
-    pathRet /= "Library/Application Support";
-    TryCreateDirectory(pathRet);
-    return pathRet / "DashData";
-#else
-    // Unix
-    return pathRet / ".dash-data";
-#endif
-#endif
-}
+boost::filesystem::path GetDataDirectory();
 
 class CDriveFile
 {
